Use waitpid() on systems that have it.
For systems that don't, we define a fakewaitpid() function that mostly
behaves the same as the real one.  The only difference, and the real
waitpid() is the one that introduces the change in behavior, is that
real waitpid() won't remove any pids from $&apids prematurely, while
fakewaitpid() (like the current es) will.  This is a byproduct of
using wait() when you're really only interested in a single pid.