Migrate to the lemon parser generator.
In theory this one has a design that's more amenable to GC-able parser
state than bison or POSIX yacc do.  The idea would be to create a
`scanparser` function either in a `%code` block in parse.y or in
`lempar.c` which would go through the parser's stack and do a GC of any
of the Tree*s or char*s it finds there.  That hasn't been implemented
yet, though.

Currently "mostly" works.  The obvious problems are that it generates
wrong exceptions on syntax errors, and it doesn't correctly "eat
through" the rest of the line on a syntax error.  Also, it crashes on
trip, but I haven't looked into that.