Rough demo of loadable primitives.
This adds a prim-loadable.c file which defines a $&getcwd primitive, and
it also adds a $&loadprims primitive which allows that $&getcwd
primitive to be loaded during shell runtime.  The code should all be
nicely portable, though I'm not sure the same can be said about all the
details with building and linking (e.g., -rdynamic).

This is only a demo -- there are a looot of loose ends.
 - What symbols _should_ actually be exposed to libraries?
 - Should we pursue librarification before doing this, so that just the
   library symbols are exported?
 - What's the best behavior for re-defining primitives?
 - How do we coordinate between shell versions and "addon" versions?
 - How would we handle it if the function signature of primitives
   changes?
 - Should some of the stuff currently in the shell be pulled out?
 - How should these get integrated into the current build setup?